feat(test): 重整singledo目录测试文件结构
将4个测试文件重组为1个标准命名的测试文件：
- 删除竞态条件测试文件: forced_race_test.go, double_check_locking_test.go, race_condition_test.go
- 合并到: singledo_test.go (单例执行模式测试)
- 保持97.4%测试覆盖率和24个测试函数
- 涵盖单例模式、并发安全、双重检查锁定功能
